organized file structure, created a blog folder, renamed application controller to blog controller, made a blog folder in the controllers which will hold the posts controller and the pages controller, and routed accordingly
