--- app/controllers/application_controller.rb
@@ -1,2 +1,0 @@
-class ApplicationController < ActionController::Base
-end
