+++ app/controllers/blog/pages_controller.rb
@@ -1,0 +1,10 @@
+module Blog
+
+  class PagesController < BlogController
+
+    def wyncode
+    end
+
+  end
+
+end
