+++ app/controllers/blog_controller.rb
@@ -1,0 +1,2 @@
+class BlogController < ActionController::Base
+end
