--- app/controllers/pages_controller.rb
@@ -1,4 +1,0 @@
-class PagesController < ApplicationController
-  def wyncode
-  end
-end
