+++ app/views/layouts/blog.html.erb
